LangGraphSugAgent/requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-core==0.3.15
requests==2.32.3
pydantic==2.10.2
typing-extensions==4.12.2
gitpython==3.1.43
python-dotenv==1.0.1
colorama==0.4.6
rich==13.9.4
mcp==1.12.4
lark-oapi==1.0.37